fixed macos compatibiliy issues arising from assuming `size_t`, `uint64_t`, and `VkDeviceSize` are the same type

+++ src/runtime/vk/CommandBuffer.cpp
@@ -953,8 +953,8 @@ namespace vuk {
 
 		auto sets_mask = sets_to_bind.to_ulong();
 		auto persistent_sets_mask = persistent_sets_to_bind.to_ulong();
-		uint64_t highest_undisturbed_binding_required = 0;
-		uint64_t lowest_disturbed_binding = VUK_MAX_SETS;
+		size_t highest_undisturbed_binding_required = 0;
+		size_t lowest_disturbed_binding = VUK_MAX_SETS;
 		for (size_t set_index = 0; set_index < VUK_MAX_SETS; set_index++) {
 			bool set_to_bind = sets_mask & (1ULL << set_index);
 			bool persistent_set_to_bind = persistent_sets_mask & (1ULL << set_index);
